+++ Resource/ModuleName.build.ps1
@@ -1,3 +1,9 @@
-$module = Import-Module Plumber -PassThru
-Get-ChildItem "$($module.ModuleBase)\Tasks" -Recurse -File |
-    ForEach-Object {. $_.FullName}
+$module = Get-Module Plumber
+if (-not $module) {
+    $module = Import-Module Plumber -PassThru
+}
+
+. (Get-PlumberTaskLoader) -Config @{
+    # Required, case sensitive name of the psd1 file of the module
+    ModuleManifest = 'ModuleName.psd1'
+}